Remove logging from filtered list views when called by `noms-ops` app because we have sufficient statistics collected
