+++ mtp_api/apps/credit/views.py
@@ -19,7 +19,6 @@ from core.filters import (
     BaseFilterSet,
     BlankStringFilter,
     IsoDateTimeFilter,
-    LogNomsOpsSearchDjangoFilterBackend,
     MultipleFieldCharFilter,
     MultipleValueFilter,
     PostcodeFilter,
@@ -268,7 +267,7 @@ class CreditViewMixin:
 
 
 class GetCredits(CreditViewMixin, mixins.ListModelMixin, viewsets.GenericViewSet):
-    filter_backends = (LogNomsOpsSearchDjangoFilterBackend, SafeOrderingFilter)
+    filter_backends = (DjangoFilterBackend, SafeOrderingFilter)
     filterset_class = CreditListFilter
     ordering_fields = ('created', 'received_at', 'amount',
                        'prisoner_number', 'prisoner_name')
